first pass at Action/Reduce and State graph
* implements a barebones framework for writing game logic as a set of
  Actions+Rules
* implements a model for state modifications as a graph structure with
  strict storage constraints a collection of Nodes+Attributes+Edges

Actions are the user commitments.

Conceptually they are the log of user intent, practically they a just an
immutable id and definition of arguments (like a function signature).

We define them in solidity as a contract that implements the
`ActionType` interface:

Rules apply `Actions` to modify `State`.

The logic of what state modifications to make for any given `Action` is
through a pure-like `reduce` function. Many Rules may operate on many
Actions.

We define then in solidity as a contract that implements the `Rule`
interface:

Dispatchers are the entrypoint to games. They define the set of `Rules`
that will be applied in order, the State to use, and the `dispatch`
function entrypoint to apply incoming `Actions`.

We define them in solidity as a contract that implements the
`Dispatcher` interface. A `BaseDispatcher` abstract contract can be used
to avoid some boilerplate:

To ensure that there is a common way for Rules to modify the game state,
and ensure we have a consistent way to index that state we perform all
state modifications through a common `State` type. Think of it like a
really simple database or key/value store.

State is a modelled as sparse directed graph of Nodes, Attributes and
Edges. `State` is an interface:

`StateGraph` is an implementation of that interface that aims to
minimise the amount of storage writes by limiting the amount of
attribute storage per NodeType to a single uint256 slot.

`Nodes` in the State are basically just an immutable ID. But we define
them in solidity as something that implements the `NodeType` interface
which describes how to decode the node's attribute data:

Edges are weighted directional relationships between Nodes.

We define them in solidity as a contract that implements the `EdgeType`
interface:

The idea of the State model is that by modelling in this manor we can
boil down all state changes to either SetNode(id, data) or SetEdge(from,
to, weight). Making it simpler to index the data.

The idea of the Action/Rules model is that by seperating the commitment
of Actions from the execution to replay and apply the state changes, we
can have a more composable architecture that lends itself to off-chain
computation.