Routers are basically all SessionRouters
The Router interface was previously light, but since Routers are so key
to cog-services and sessionkeys it doesn't really make much sense, so
bring in the auth funcs to IRouter.

also removes the "BasicGame" helper, replaces with BaseGame abstract
class. BasicGame did too much of the setup for you that you couldn't
actually use it properly. This brings us closer to something updatable.