add type hints for weight value
we do not only care about weights connected to certain nodes. when
calculating aggregate values, we need to know how to decode all the
weights. so the idea of special "attribute" node types is very limiting.

instead we can annouce type hints for the weight value when we register
the edge type. unfortunatly this is not enforced anywhere, but at least
it allows the indexer to make suitable judgements about the shape of
weight data.