try "fat-nodes" instead of "fat-edges"
experimenting with creating a larger node id space (large enough to fit
an address in) instead of a large "weight" value.

this allows us to have nodes that represent EOA accounts or contract
addresses.

but at the cost of now only having 64bits for the edge weight.

since we were using edge weights for "properties" this may be limiting

however for certain kinds of data, it may be reasonable to abuse the
node-id storage in the outbound edge data by creating a node type