annotations: references to off-chain state
sometimes we want to commit or reference a large blob of data where it
is very important that clients know what the data is, but not important
that on-chain code know what the data is.

for these cases, we can make use of the transaction calldata to store
some bytes, which the client can fetch.

however we then need to pass a reference to that data to on-chain logic,
but how can we be sure that the reference we were given corrosponds to
data that has been made available to clients.

annotations are a solution to this problem.

an annotation is a label attached to a Node that points to a hash of
some data that is guarenteed to have been made available in calldata.

when a user calls "dispatch", they can optionally attach one or more
blobs of data. these data are then hashed to create annotation
references, and these refs are passed along with the action call context.
actions can then annotate nodes in the state graph with this ref and
know for sure that the client can find the full data value for the
annotation by querying the calldata.

You can think of annotations a bit like how email attachments work. You
add your attachments to your dispatch, and then refernce the attachments
by id in action rules.

Some examples where this is useful:

* Annotating text descriptions on things
* Storing blob of JSON metadata
* Publishing image assets