simpler annotations
* removes "attachments" style
* only support UTF8 string data (99% use case)

the "attachments" idea for annotations (where you bundle annotation
attachments along with dispatch calls) was intended to remove the need
to convert the large calldata bytes to large memory bytes. but this
complicated both the dispatch and annotation apis and had very little
benfits (since kecack casts to memory anyway). So this is dropped and
annotations are now just a State thing, no polluting of the Dispatcher

originally annotations were to be opaque blobs of bytes. but for this to
be useful we really need some kind of mime-type support to allow the
indexers/clients to make informed decisions about decoding. instead we
are now saying all annotations are UTF8 strings... they might be strings
containing JSON, or base64 strings, or complex data URLs, but they are
always a UTF8 string - this makes the indexers job easier and easier for
humans to read.