support subscribing to Edge/Annotation events
subscribing to large State queries is bad for performance. I removed
this feature as it was a footgun.

instead we should subscribe to the edge/annotation changes and then
let the client work out what it needs to re-query.

bonus; fixes the matcher bug that dropped edges in match results by
rewriting the match logic in a more naive way ... when it's time for
persistance we will not want to be using any of this logic anyway and
hopefully use some more full featured lib/db